refactor(apigatewayv2): split unit and integration tests per naming convention
Separate host-parsing unit tests into WebSocketHandlerTest and rename the
integration test to WebSocketSubdomainConnectIntegrationTest, aligning with
the project's *Test / *IntegrationTest naming convention.
